Add GET /api/me session check; rename LogoutHandler -> SessionEndpointHandler
Browsers can't read the status of a rejected WebSocket upgrade, so a cookie-auth
client can't tell "not signed in" from "server down" and ends up retrying an
unauthenticated socket forever. Add a readable REST probe: GET /api/me returns 204
when the session cookie authenticates the caller, 401 otherwise (a session read —
the /api/* SessionHandler has already resolved the cookie).

Folded into the former LogoutHandler since both are browser session-lifecycle
routes; renamed to SessionEndpointHandler to avoid clashing with Vert.x's own
SessionHandler in ApiGatewayVertcleFactory. Stage 1 of the /api/me work; the core
client pre-flight and frontend wiring follow.
